backend/requirements.txt
# Core FastAPI & Server
fastapi==0.126.0
uvicorn[standard]==0.40.0
python-dotenv==1.2.1

# AI Integration (Corrected Library)
google-genai==1.56.0

# Database & Migrations
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
alembic==1.17.2

# Data Validation & Parsing
pydantic==2.12.5
email-validator==2.3.0

# File Handling & PDF
python-multipart==0.0.21
PyMuPDF==1.26.7
reportlab==4.4.7

# Authentication & Security
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
cryptography==46.0.3

# Utilities
httpx==0.28.1
requests==2.32.5